Refactored both Tooltip and Popover to use AbstractTooltip which contains common code for both.
